requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.2
Flask==1.1.2  
Flask-WTF==1.0.0   
WTForms==3.0.0 
Werkzeug==1.0.1 
Jinja2==2.11.3
numpy==1.20.1  
MarkupSafe==1.1.1  
itsdangerous==1.1.0  
pillow==8.2.0  
scikit-learn==0.24.1